add hooks to feature box hidden content
These hooks display above and below the widget areas. In order to do so
without breaking the layout, the widget areas got a new <div> wrapper.

Now with the hooks in place, the hidden content conditional was
adjusted to allow for the .feature-box-hidden <div> to display if there
are active widget areas OR if the forced toggle option is checked. This
means that you can have no widgets in place but force the toggle button
and use the hooks to build your own hidden content… and the toggle will
still work as the JS still uses the same HTML elements.